rename modules and fix logging messages
