+++ sdk/eventhub/event-hubs/src/batchingPartitionChannel.ts
@@ -13,7 +13,7 @@ import { isDefined, isObjectWithProperties } from "@azure/core-util";
 import { AbortSignalLike } from "@azure/abort-controller";
 import { AwaitableQueue } from "./impl/awaitableQueue";
 import { getPromiseParts } from "./util/getPromiseParts";
-import { logger } from "./log";
+import { logger } from "./logger";
 
 export interface BatchingPartitionChannelProps {
   loopAbortSignal: AbortSignalLike;
