+++ sdk/eventhub/event-hubs/src/util/error.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { logErrorStackTrace, logger } from "../log";
+import { logErrorStackTrace, logger } from "../logger";
 import { ConnectionContext } from "../connectionContext";
 import { isDefined } from "@azure/core-util";
 import { AmqpError, isAmqpError } from "rhea-promise";
